DPI handling, ReadOnly config file handling
-Server and Client now are monitor DPI aware. User can adjust how much text scaling should occurs to compensate for small text on high resolution screens
-Was unable to save config when file was set to 'readonly'. This situation could happen when file is submitted to a source control. Now create a second _user config file, if default one is readonly
-Fixed crash when closing Server application. Tied to Server Listening thread with socket never timeout and unable to exit

+++ Code/Client/Private/NetImgui_CmdPackets_DrawFrame.cpp
@@ -189,7 +189,7 @@ CmdDrawFrame* CompressCmdDrawFrame(const CmdDrawFrame* pDrawFramePrev, const Cmd
 	//-----------------------------------------------------------------------------------------
 	// Allocate memory for worst case scenario (no compression possible)
 	// New DrawFrame size + 2 'compression block info' per data stream
-	size_t neededDataCount			= DivUp<size_t>(pDrawFrameNew->mHeader.mSize, ComDataSize) + 6*pDrawFrameNew->mDrawGroupCount;
+	size_t neededDataCount			= DivUp<size_t>(pDrawFrameNew->mHeader.mSize, ComDataSize) + 6*static_cast<size_t>(pDrawFrameNew->mDrawGroupCount);
 	CmdDrawFrame* pDrawFramePacked	= netImguiSizedNew<CmdDrawFrame>(neededDataCount*ComDataSize);	
 	*pDrawFramePacked				= *pDrawFrameNew;
 	pDrawFramePacked->mCompressed	= true;
